Factor out cookie validation so it can be shared by ProxyRoutesSpec as well
