+++ src/test/scala/org/broadinstitute/dsde/workbench/leonardo/api/TestLeoRoutes.scala
@@ -2,7 +2,7 @@ package org.broadinstitute.dsde.workbench.leonardo.api
 
 import java.io.ByteArrayInputStream
 
-import akka.http.scaladsl.model.headers.OAuth2BearerToken
+import akka.http.scaladsl.model.headers.{HttpCookiePair, OAuth2BearerToken, `Set-Cookie`}
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 import com.typesafe.config.ConfigFactory
 import net.ceedubs.ficus.Ficus._
@@ -17,15 +17,15 @@ import org.broadinstitute.dsde.workbench.leonardo.monitor.NoopActor
 import org.broadinstitute.dsde.workbench.leonardo.service.{BucketHelper, LeonardoService, MockProxyService, StatusService}
 import org.broadinstitute.dsde.workbench.model.google._
 import org.broadinstitute.dsde.workbench.model.{UserInfo, WorkbenchEmail, WorkbenchUserId}
+import org.scalatest.Matchers
 import org.scalatest.concurrent.ScalaFutures
 
-import scala.concurrent.Future
 import scala.concurrent.duration._
 
 /**
   * Created by rtitle on 8/15/17.
   */
-trait TestLeoRoutes { this: ScalatestRouteTest with ScalaFutures =>
+trait TestLeoRoutes { this: ScalatestRouteTest with ScalaFutures with Matchers =>
   val config = ConfigFactory.parseResources("reference.conf").withFallback(ConfigFactory.load())
   val whitelist = config.as[Set[String]]("auth.whitelistProviderConfig.whitelist").map(_.toLowerCase)
   val dataprocConfig = config.as[DataprocConfig]("dataproc")
@@ -61,6 +61,9 @@ trait TestLeoRoutes { this: ScalatestRouteTest with ScalaFutures =>
   val statusService = new StatusService(mockGoogleDataprocDAO, mockSamDAO, DbSingleton.ref, dataprocConfig, pollInterval = 1.second)
   val defaultUserInfo = UserInfo(OAuth2BearerToken("accessToken"), WorkbenchUserId("user1"), WorkbenchEmail("user1@example.com"), 0)
   val tokenAge = 500000
+  val tokenName = "LeoToken"
+  val tokenValue = "accessToken"
+  val tokenCookie = HttpCookiePair(tokenName, tokenValue)
   val timedUserInfo = defaultUserInfo.copy(tokenExpiresIn = tokenAge)
   val leoRoutes = new LeoRoutes(leonardoService, proxyService, statusService, swaggerConfig) with MockUserInfoDirectives {
     override val userInfo: UserInfo = defaultUserInfo
@@ -75,5 +78,20 @@ trait TestLeoRoutes { this: ScalatestRouteTest with ScalaFutures =>
 
   def notebookServiceAccount(googleProject: GoogleProject): Option[WorkbenchEmail] = {
     serviceAccountProvider.getNotebookServiceAccount(defaultUserInfo, googleProject).futureValue
+  }
+
+  private[api] def validateCookie(setCookie: Option[`Set-Cookie`],
+                             expectedCookie: HttpCookiePair = tokenCookie,
+                             age: Long = tokenAge / 1000): Unit = {
+    def roundUpToNearestTen(d: Long) = Math.ceil(d / 10.0) * 10
+
+    setCookie shouldBe 'defined
+    val cookie = setCookie.get.cookie
+    cookie.name shouldBe expectedCookie.name
+    cookie.value shouldBe expectedCookie.value
+    cookie.secure shouldBe true
+    cookie.maxAge.map(roundUpToNearestTen) shouldBe Some(age) // test execution loses some milliseconds
+    cookie.domain shouldBe None
+    cookie.path shouldBe Some("/")
   }
 }
